+++ scripts/build-wasm.ps1
@@ -3,7 +3,9 @@
 $ErrorActionPreference = 'Stop'
 $PSNativeCommandUseErrorActionPreference = $true
 
-rm -Recurse -Force rust/kcl-wasm-lib/pkg
+if (Test-Path rust/kcl-wasm-lib/pkg) {
+    rm -Recurse -Force rust/kcl-wasm-lib/pkg
+}
 mkdir -p rust/kcl-wasm-lib/pkg
 if (Test-Path rust/kcl-lib/bindings) {
     rm -Recurse -Force rust/kcl-lib/bindings
